IDEA-163051 "Replace inefficient Stream API call chains ending with count()": fix leaves redundant casts IDEA-163050 "Replace inefficient Stream API call chains ending with count()": highlighting
(cherry picked from commit c8b125f; IDEA-CR-14950)

+++ java/java-tests/testData/inspection/inefficientStreamCount/beforeStreamCountCastToInt.java
@@ -3,7 +3,7 @@
 import java.util.Arrays;
 
 class Test {
-  long cnt() {
-    return Arrays.asList('d', 'e', 'f').stream().c<caret>ount();
+  int cnt() {
+    return (int) Arrays.asList('d', 'e', 'f').stream().c<caret>ount();
   }
 }
